refactor(domain): 수익 계산하는 메서드 BetAmounts로 이동
